Sat, Jan 7, 2023, 5:45 PM +00:00

+++ filter-more/helpers.c
@@ -73,27 +73,27 @@ for (int a = 1; a < height - 1; a++)
     {
         for (int b = 1; b < width - 1 ; b++)
         {
-            int GX_red = -1*(copy[a-1][b-1].rgbtRed) + 0*(copy[a-1][b].rgbtRed) + 1*(copy[a-1][b+1].rgbtRed)
+            BYTE GX_red = -1*(copy[a-1][b-1].rgbtRed) + 0*(copy[a-1][b].rgbtRed) + 1*(copy[a-1][b+1].rgbtRed)
             + (-2*(copy[a][b-1].rgbtRed)) + 0*(copy[a][b].rgbtRed) + 2*(copy[a][b+1].rgbtRed)
             + (-1*(copy[a+1][b-1].rgbtRed)) + 0*(copy[a+1][b].rgbtRed) + 1*(copy[a+1][b+1].rgbtRed);
 
-            int GX_green = -1*(copy[a-1][b-1].rgbtGreen) + 0*(copy[a-1][b].rgbtGreen) + 1*(copy[a-1][b+1].rgbtGreen)
+            BYTE GX_green = -1*(copy[a-1][b-1].rgbtGreen) + 0*(copy[a-1][b].rgbtGreen) + 1*(copy[a-1][b+1].rgbtGreen)
             + (-2*(copy[a][b-1].rgbtGreen)) + 0*(copy[a][b].rgbtGreen) + 2*(copy[a][b+1].rgbtGreen)
             + (-1*(copy[a+1][b-1].rgbtGreen)) + 0*(copy[a+1][b].rgbtGreen) + 1*(copy[a+1][b+1].rgbtGreen);
 
-           int GX_blue = -1*(copy[a-1][b-1].rgbtBlue) + 0*(copy[a-1][b].rgbtBlue) + 1*(copy[a-1][b+1].rgbtBlue)
+           BYTE GX_blue = -1*(copy[a-1][b-1].rgbtBlue) + 0*(copy[a-1][b].rgbtBlue) + 1*(copy[a-1][b+1].rgbtBlue)
             + (-2*(copy[a][b-1].rgbtBlue)) + 0*(copy[a][b].rgbtBlue) + 2*(copy[a][b+1].rgbtBlue)
             + (-1*(copy[a+1][b-1].rgbtBlue)) + 0*(copy[a+1][b].rgbtBlue) + 1*(copy[a+1][b+1].rgbtBlue);
 
-            int GY_red = -1*(copy[a-1][b-1].rgbtRed) + (-2*(copy[a-1][b].rgbtRed)) + (-1*(copy[a-1][b+1].rgbtRed))
+            BYTE GY_red = -1*(copy[a-1][b-1].rgbtRed) + (-2*(copy[a-1][b].rgbtRed)) + (-1*(copy[a-1][b+1].rgbtRed))
             + 0*(copy[a][b-1].rgbtRed) + 0*(copy[a][b].rgbtRed) + 0*(copy[a][b+1].rgbtRed)
             + 1*(copy[a+1][b-1].rgbtRed) + 2*(copy[a+1][b].rgbtRed) + 1*(copy[a+1][b+1].rgbtRed);
 
-            int GY_green = -1*(copy[a-1][b-1].rgbtGreen) + (-2*(copy[a-1][b].rgbtGreen)) + (-1*(copy[a-1][b+1].rgbtGreen))
+            BYTE GY_green = -1*(copy[a-1][b-1].rgbtGreen) + (-2*(copy[a-1][b].rgbtGreen)) + (-1*(copy[a-1][b+1].rgbtGreen))
             + 0*(copy[a][b-1].rgbtGreen) + 0*(copy[a][b].rgbtGreen) + 0*(copy[a][b+1].rgbtGreen)
             + 1*(copy[a+1][b-1].rgbtGreen) + 2*(copy[a+1][b].rgbtGreen) + 1*(copy[a+1][b+1].rgbtGreen);
 
-            int GY_blue = -1*(copy[a-1][b-1].rgbtBlue) + (-2*(copy[a-1][b].rgbtBlue)) + (-1*(copy[a-1][b+1].rgbtBlue))
+            BYTE GY_blue = -1*(copy[a-1][b-1].rgbtBlue) + (-2*(copy[a-1][b].rgbtBlue)) + (-1*(copy[a-1][b+1].rgbtBlue))
             + 0*(copy[a][b-1].rgbtBlue) + 0*(copy[a][b].rgbtBlue) + 0*(copy[a][b+1].rgbtBlue)
             + 1*(copy[a+1][b-1].rgbtBlue) + 2*(copy[a+1][b].rgbtBlue) + 1*(copy[a+1][b+1].rgbtBlue);
